requirements.txt
dnspython==1.16.0
Flask==2.0.3
Flask-PyMongo==2.3.0
Flask-QRcode==3.0.0
Flask-WTF==1.0.1
python-dotenv==0.20.0